Started set-dressing and added context to "the orb"
Began work on sprucing up the scene a bit by adding a table and base for the orb.
Also imported a few more assets to be used in set dressing.
Whilst adding a glass-ball around the level, it seems like the additive shader stops working, need to look into that.